scripts/filter_similar_images/requirements.txt
facenet_sandberg>=0.0.8
dhash>=1.3
numpy>=1.14.2
typing>=3.6.4
Pillow>=5.3.0
